MET-2877 Convert criteria containers to filters

+++ metis-indexing/src/main/java/eu/europeana/indexing/IndexerImpl.java
@@ -125,11 +125,11 @@ class IndexerImpl implements Indexer {
 
   @Override
   public void index(InputStream record, Date recordDate,
-          boolean preserveUpdateAndCreateTimesFromRdf, List<String> datasetIdsForRedirection,
-          boolean performRedirects) throws IndexingException {
+      boolean preserveUpdateAndCreateTimesFromRdf, List<String> datasetIdsForRedirection,
+      boolean performRedirects) throws IndexingException {
     final StringToFullBeanConverter stringToRdfConverter = stringToRdfConverterSupplier.get();
     indexRdf(stringToRdfConverter.convertToRdf(record), recordDate,
-            preserveUpdateAndCreateTimesFromRdf, datasetIdsForRedirection, performRedirects);
+        preserveUpdateAndCreateTimesFromRdf, datasetIdsForRedirection, performRedirects);
   }
 
   @Override
@@ -155,7 +155,9 @@ class IndexerImpl implements Indexer {
   @Override
   public int removeAll(String datasetId, Date maxRecordDate)
       throws IndexerRelatedIndexingException {
-    return this.connectionProvider.getIndexedRecordAccess().removeDataset(datasetId, maxRecordDate);
+    // TODO: 8/26/20 Update removeAll method to return long instead of int, it will affect clients
+    return Math.toIntExact(
+        this.connectionProvider.getIndexedRecordAccess().removeDataset(datasetId, maxRecordDate));
   }
 
   @Override
